refactor: 일정, 발자국 패키지 분리
