+++ backend/src/main/java/kr/ac/hs/RandomTrip/footprint/dto/FootprintRequestDto.java
@@ -1,0 +1,12 @@
+package kr.ac.hs.RandomTrip.footprint.dto;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class FootprintRequestDto {
+    private Long destinationId;
+    private String memo;
+    private String photoUrl;
+}
