+++ backend/src/main/java/kr/ac/hs/RandomTrip/itinerary/dto/ItineraryItemOrderRequestDto.java
@@ -1,0 +1,12 @@
+package kr.ac.hs.RandomTrip.itinerary.dto;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Getter
+@NoArgsConstructor
+public class ItineraryItemOrderRequestDto {
+    private List<Long> itemIds;
+}
