+++ backend/src/main/java/kr/ac/hs/RandomTrip/itinerary/dto/ItineraryItemRequestDto.java
@@ -1,0 +1,12 @@
+package kr.ac.hs.RandomTrip.itinerary.dto;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class ItineraryItemRequestDto {
+    private Long destinationId;
+}
