+++ backend/src/main/java/kr/ac/hs/RandomTrip/itinerary/dto/ItineraryRequestDto.java
@@ -1,0 +1,10 @@
+package kr.ac.hs.RandomTrip.itinerary.dto;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class ItineraryRequestDto {
+    private String name;
+}
