--- backend/src/main/java/kr/ac/hs/RandomTrip/trip/domain/Itinerary.java
@@ -1,45 +1,0 @@
-package kr.ac.hs.RandomTrip.trip.domain;
-
-import jakarta.persistence.*;
-import kr.ac.hs.RandomTrip.auth.domain.Member;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-@EntityListeners(AuditingEntityListener.class)
-public class Itinerary {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String name;
-
-    @CreatedDate
-    @Column(updatable = false)
-    private LocalDateTime createdAt;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
-
-    @OneToMany(mappedBy = "itinerary", cascade = CascadeType.ALL, orphanRemoval = true)
-    @OrderBy("itemOrder ASC")
-    private List<ItineraryItem> items = new ArrayList<>();
-
-    @Builder
-    public Itinerary(String name, Member member) {
-        this.name = name;
-        this.member = member;
-    }
-}
