--- backend/src/main/java/kr/ac/hs/RandomTrip/trip/dto/footprint/FootprintRequestDto.java
@@ -1,12 +1,0 @@
-package kr.ac.hs.RandomTrip.trip.dto.footprint;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class FootprintRequestDto {
-    private Long destinationId;
-    private String memo;
-    private String photoUrl;
-}
