--- backend/src/main/java/kr/ac/hs/RandomTrip/trip/dto/itinerary/ItineraryItemOrderRequestDto.java
@@ -1,12 +1,0 @@
-package kr.ac.hs.RandomTrip.trip.dto.itinerary;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Getter
-@NoArgsConstructor
-public class ItineraryItemOrderRequestDto {
-    private List<Long> itemIds;
-}
