--- backend/src/main/java/kr/ac/hs/RandomTrip/trip/dto/itinerary/ItineraryItemRequestDto.java
@@ -1,12 +1,0 @@
-package kr.ac.hs.RandomTrip.trip.dto.itinerary;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-public class ItineraryItemRequestDto {
-    private Long destinationId;
-}
