--- backend/src/main/java/kr/ac/hs/RandomTrip/trip/dto/itinerary/ItineraryRequestDto.java
@@ -1,10 +1,0 @@
-package kr.ac.hs.RandomTrip.trip.dto.itinerary;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class ItineraryRequestDto {
-    private String name;
-}
