+++ backend/src/main/java/kr/ac/hs/RandomTrip/trip/service/TourService.java
@@ -5,9 +5,10 @@ import kr.ac.hs.RandomTrip.auth.repository.MemberRepository;
 import kr.ac.hs.RandomTrip.trip.domain.Destination;
 import kr.ac.hs.RandomTrip.trip.domain.Visit;
 import kr.ac.hs.RandomTrip.trip.dto.GuideResponse;
+import kr.ac.hs.RandomTrip.trip.llm.GeminiGuideGenerator;
 import kr.ac.hs.RandomTrip.trip.repository.DestinationRepository;
-import kr.ac.hs.RandomTrip.trip.domain.ItineraryItem;
-import kr.ac.hs.RandomTrip.trip.repository.ItineraryItemRepository;
+import kr.ac.hs.RandomTrip.itinerary.domain.ItineraryItem;
+import kr.ac.hs.RandomTrip.itinerary.repository.ItineraryItemRepository;
 import kr.ac.hs.RandomTrip.trip.repository.VisitRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.context.SecurityContextHolder;
